[SPARK-56505][SQL][TEST] Add SessionQueryTest to replace SharedSparkSession
### What changes were proposed in this pull request?

This PR adds `SessionQueryTest` as the new, connect-compatible test base trait to facilitate
code-reuse of sql/core tests in sql/connect by making it possible to add 'connect variants'.

If some `FooSuite` in `sql/core` uses the new `sql.SessionQueryTest` trait like e.g.

```
class FooSuite extends SessionQueryTest {
  checkAnswer(
    sql("SELECT 1"),
    Seq(1))
}
```

We can now add a connect variant of that suite in `sql/connect` as follows:

```
class FooWithConnectSuite extends FooSuite with connect.SessionQueryTest
```

To make this possible, the PR makes the following structural changes:

- **Factor the test helpers out of `QueryTest`** into focused, session-agnostic traits:
  - `CheckAnswerHelper` (sql/core) - `checkAnswer` and friends.
  - `QueryCleanupHelper` (sql/core) - `withTable` / `withView` / `withUserDefinedFunction`.
  - `RowComparisonUtils` (sql/core) - the pure row-comparison logic (`prepareAnswer`,
    `prepareRow`, `genError`, `compare`, `sameRows`); the legacy `object QueryTest` now delegates
    to it instead of keeping a forked copy.
  - `CheckErrorHelper` (core, extracted from `SparkTestSuite`) - the `checkError` helpers.
- **Introduce a session-injection seam** so the bound `SparkSession` can be specialized per module:
  - `SparkSessionProvider` exposes an abstract `spark`.
  - `SparkSessionBinder` provides the concrete classic session and its lifecycle (the thread
    audit, `beforeAll`/`afterAll`, `DebugFilesystem` checks).
  - classic (`classic.SparkSessionBinder`) and connect (`connect.SparkSessionBinder`) leaves bind
    the respective session types; `connect.SessionQueryTest` overrides `spark` to a
    `connect.SparkSession` and overrides `isDfSorted` to inspect the explain string.
- **Add `SessionQueryTestBase` / `sql.SessionQueryTest` / `classic.SessionQueryTest` /
  `connect.SessionQueryTest`** as the trait hierarchy suites pick from. A binary `isConnect:
  Boolean` discriminator (`false` for classic, `true` for connect) lets a shared test body
  document session-specific behaviour via `if (isConnect) ... else ...`.
- **Rebase `SharedSparkSession` onto the new binder traits.** `SharedSparkSession` becomes a thin
  trait over `classic.SparkSessionBinder` (preserving the existing thread-audit behaviour), and
  `SharedSparkSessionBase` over `SparkSessionBinderBase`; behaviour is unchanged. `SparkSessionBinderBase`
  is retained for now because `SharedSparkSessionBase` is still used by suites such as
  `GenericWordSpecSuite` (TODO to migrate and remove it).
- **Add the `ExampleSessionAgnosticSuite` example suites** (colocated in sql/connect) demonstrating
  the agnostic style and the `withConf` helper (a public-API-only alternative to `withSQLConf`),
  plus the `QueryTestWithConnectSuite` connect variant of `QueryTestSuite`. `ParquetQuerySuite` and
  `QueryTestSuite` are rebased onto the new traits.

The deprecations of the now-superseded entry points (`SharedSparkSession`, the `QueryTest` object,
`checkAggregatesWithTol`, `stripSparkFilter`, `logicalPlanToSparkQuery`, `makeQualifiedPath`,
`withCurrentCatalogAndNamespace`) are intentionally deferred to a follow-up PR.

Class hierarchy (sql/core unless noted):

```
SessionQueryTestBase  (AnyFunSuite + SparkSessionProvider + CheckAnswerHelper
                       + CheckErrorHelper + QueryCleanupHelper; abstract isConnect)
  + sql.SessionQueryTest        (SparkFunSuite + SparkSessionBinder; isConnect = false)
      + classic.SessionQueryTest    (+ classic.SparkSessionBinder)
      + connect.SessionQueryTest    (connect, + connect.SparkSessionBinder; isConnect = true)
```

c.f. https://github.com/apache/spark/pull/56190#discussion_r3359763783 for more background and
motivation.

### Why are the changes needed?

Currently, most tests use `SharedSparkSession` to obtain the `spark` object. This prevents
specializing these tests in `sql/connect` as `SharedSparkSession` provides a `classic.SparkSession`,
thus preventing overriding.

### Does this PR introduce _any_ user-facing change?

No. This patch is test-only. `SharedSparkSession`'s behaviour is preserved (it is rebased onto the
new binder traits without changing semantics).

### How was this patch tested?

This patch is test-only. Existing suites (`QueryTestSuite`, `ParquetQuerySuite`, ...) are rebased
onto the new traits and continue to pass; new example/connect-variant suites
(`ExampleSessionAgnosticSuite`, `QueryTestWithConnectSuite`) exercise the new agnostic and connect
paths.

### Was this patch authored or co-authored using generative AI tooling?

Parts of this patch were authored by claude code.

Closes #56190 from fwc/sharedsparksession-refactor-mostly-nonbreaking.

Lead-authored-by: Matthis Gördel <29337229+fwc@users.noreply.github.com>
